+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - API
     - Devices
       - [HMIP-HDM1] (Hunter Douglas & erfal window blinds)
+      - [HMIP-HAP] (HomematicIP Access Point)
 
 
 ### CHANGED
@@ -380,3 +381,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [HMIP-FSI16]: https://github.com/coreGreenberet/homematicip-rest-api/issues/310
 [HMIP-SWDO-PL]: https://github.com/coreGreenberet/homematicip-rest-api/issues/315
 [HMIP-HDM1]: https://github.com/coreGreenberet/homematicip-rest-api/issues/332
+[HMIP-HAP]: https://github.com/coreGreenberet/homematicip-rest-api/issues/335